SUN-76: Refine test architecture
